Actualizado testing strategy * Pages no debe contener acciones * Usando getByRole en homepage test heading title * ContentList actions con test.step
