+++ tests/actions/ContentListActions.ts
@@ -1,25 +1,17 @@
-import { getSectionRoute } from '@utils/sections'
+import { getURLForSection } from '@utils/sections'
 import type { SectionType } from '@domain/section'
 import type { UILanguages } from '@i18n/ui'
-import type { ContentListPage } from '@tests/pages/ContentListPage'
-import type { Page } from '@playwright/test'
+import { test, type Page } from '@playwright/test'
 
-export async function goto(page: Page, locale: UILanguages, collection: SectionType) {
-  const route = getSectionRoute(collection, locale)
-  await page.goto(`/${locale}/${route}`)
+export async function openCollectionList(page: Page, locale: UILanguages, collection: SectionType) {
+  await test.step(`Open ${collection} list in ${locale}`, async () => {
+    await page.goto(getURLForSection(collection, locale))
+  })
 }
 
-export function getItemPath(locale: UILanguages, collection: SectionType, itemPath: string) {
-  const route = getSectionRoute(collection, locale)
-  const href = `/${locale}/${route}/${itemPath}`
-  const pattern = `${locale}/${route}/`
-  return { pattern, href }
-}
-
-export async function openItem(list: ContentListPage, locale: UILanguages, collection: SectionType, itemPath: string) {
-  const { pattern, href } = getItemPath(locale, collection, itemPath)
-  await Promise.all([
-    list.page.waitForURL(new RegExp(pattern)),
-    list.page.locator(`[href="${href}"]`).click()
-  ])
+export async function clickLinkItem(page: Page, locale: UILanguages, collection: SectionType, itemPath: string) {
+  await test.step(`Open ${collection} item ${itemPath} in ${locale}`, async () => {
+    const href = `${getURLForSection(collection, locale)}/${itemPath}`
+    await page.locator(`[href="${href}"]`).click()
+  })
 }
