+++ tests/pages/BlogPage.ts
@@ -7,22 +7,11 @@ export class BlogPage {
     this.page = page
   }
 
-  async gotoList(path: string) {
-    await this.page.goto(path)
-  }
-
   headerTitle(): Locator {
     return this.page.getByTestId('header-title')
   }
 
   tagLink(tag: string): Locator {
     return this.page.getByTestId(`tag-link-${tag}`)
-  }
-
-  async openPost(href: string, locale: string) {
-    await Promise.all([
-      this.page.waitForURL(new RegExp(`${locale}/blog/`)),
-      this.page.locator(`[href="${href}"]`).click()
-    ])
   }
 }
